Rename: postBoardWrite to postBoardWriteAPI (#17)
